move Documentation to docs folder
