+++ .gitignore
@@ -14,5 +14,5 @@ benchmark.txt
 !/.travis.yml
 !/.promu.yml
 !/.golangci.yml
-/documentation/examples/remote_storage/remote_storage_adapter/remote_storage_adapter
-/documentation/examples/remote_storage/example_write_adapter/example_writer_adapter
+/docs/examples/remote_storage/remote_storage_adapter/remote_storage_adapter
+/docs/examples/remote_storage/example_write_adapter/example_writer_adapter
