+++ docs/examples/custom-sd/adapter/adapter.go
