+++ docs/examples/custom-sd/adapter/adapter_test.go
