+++ docs/examples/kubernetes-rabbitmq/README.md
