+++ docs/examples/kubernetes-rabbitmq/rc.yml
