+++ docs/examples/kubernetes-rabbitmq/svc.yml
