+++ docs/examples/prometheus-kubernetes.yml
