+++ docs/examples/prometheus-marathon.yml
