+++ docs/examples/prometheus.yml
