+++ docs/examples/remote_storage/example_write_adapter/README.md
