+++ docs/examples/remote_storage/example_write_adapter/server.go
