+++ docs/examples/remote_storage/remote_storage_adapter/graphite/client_test.go
