+++ docs/examples/remote_storage/remote_storage_adapter/graphite/escape.go
