+++ docs/examples/remote_storage/remote_storage_adapter/influxdb/client.go
