+++ docs/examples/remote_storage/remote_storage_adapter/main.go
@@ -40,9 +40,9 @@ import (
 	"github.com/prometheus/common/promlog"
 	"github.com/prometheus/common/promlog/flag"
 
-	"github.com/prometheus/prometheus/documentation/examples/remote_storage/remote_storage_adapter/graphite"
-	"github.com/prometheus/prometheus/documentation/examples/remote_storage/remote_storage_adapter/influxdb"
-	"github.com/prometheus/prometheus/documentation/examples/remote_storage/remote_storage_adapter/opentsdb"
+	"github.com/prometheus/prometheus/docs/examples/remote_storage/remote_storage_adapter/graphite"
+	"github.com/prometheus/prometheus/docs/examples/remote_storage/remote_storage_adapter/influxdb"
+	"github.com/prometheus/prometheus/docs/examples/remote_storage/remote_storage_adapter/opentsdb"
 	"github.com/prometheus/prometheus/prompb"
 )
