+++ docs/examples/remote_storage/remote_storage_adapter/opentsdb/client_test.go
