+++ docs/examples/remote_storage/remote_storage_adapter/opentsdb/tagvalue.go
