+++ docs/images/architecture.xml
