+++ docs/images/diagram_note.md
