+++ docs/images/internal_architecture.xml
