+++ docs/internal_architecture.md
