+++ docs/prometheus-mixin/config.libsonnet
