+++ docs/prometheus-mixin/dashboards.jsonnet
